+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Compatibility with IE11
 - `Sidebar` component added
 - Fixed: component wrapped with `withValidation` hoc didn't display an error
+- Feature: new Dropdown props (open / direction)
 
 ## 0.5.0